# openapi.yaml
openapi: 3.0.3
info:
  title: SLSEA Real-Time Solar Generation Data API
  version: 0.1.0
  description: |
    REST API for the Sri Lanka Sustainable Energy Authority (SLSEA). Rooftop solar
    meters push generation readings (write path); SLSEA users read the hierarchy,
    installations and generation history, scoped to their jurisdiction (read path).

    ## What is live now

    **Implemented:** `/health` and the whole read path: provinces, districts, grid substations
    and installations (collections with their filters, and members), the installation composite,
    the last-known reading, and one installation's readings collection and reading member.
    Also reading ingestion: `POST /installations/{installation-id}/readings` (201 with
    `Location`, `ETag` and `Last-Modified`; 409 for a duplicate). Readings are append-only:
    PUT, PATCH and DELETE on a reading or on the readings collection return 405 with `Allow`.
    Registry: `POST /installations` (201 with `Location`, `ETag` and `Last-Modified`; 409 when
    the installation or meter id is taken), `PUT /installations/{installation-id}` (whole-document
    replacement, never a merge; PATCH returns 405) and `DELETE` (200, then 404; readings are kept),
    both honouring `If-Match` (a stale ETag → 412).
    `/installations` and the readings history are paginated and sortable; installations filter by
    jurisdiction and `status`, readings by time window and `min-power-kw`.
...

...

...

...

...
    and both summaries.

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
...
paths:
...
  /provinces:
    get:
      tags: [Hierarchy]
      operationId: listProvinces
      summary: List provinces in the caller's jurisdiction
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: Provinces (flat array, not paginated)
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Vary: { $ref: '#/components/headers/Vary' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Province' }
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /provinces/{province-id}:
    parameters:
      - $ref: '#/components/parameters/ProvinceId'
    get:
      tags: [Hierarchy]
      operationId: getProvince
      summary: Get one province
      parameters:
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: The province
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Last-Modified: { $ref: '#/components/headers/LastModified' }
            Vary: { $ref: '#/components/headers/Vary' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Province' }
        '304': { $ref: '#/components/responses/NotModified' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/ProvinceNotFound' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /provinces/{province-id}/generation-summary:
    parameters:
      - $ref: '#/components/parameters/ProvinceId'
    get:
      tags: [Summaries]
      operationId: getProvinceGenerationSummary
      summary: Current total power and a day's energy across a province
      parameters:
        - $ref: '#/components/parameters/Date'
        - $ref: '#/components/parameters/IfNoneMatch'
        - $ref: '#/components/parameters/IfModifiedSince'
      responses:
        '200':
          description: The province summary
          headers:
            ETag: { $ref: '#/components/headers/ETag' }
            Vary: { $ref: '#/components/headers/Vary' }
            Cache-Control: { $ref: '#/components/headers/CacheControl' }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ProvinceGenerationSummary' }
        '304': { $ref: '#/components/responses/NotModified' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '406': { $ref: '#/components/responses/NotAcceptable' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
    post:
      tags: [Summaries]
      operationId: postProvinceGenerationSummaryNotAllowed
      summary: Not allowed. A generation summary is read-only (405)
      security: []
      responses:
        '405': { $ref: '#/components/responses/MethodNotAllowedSummary' }
    put:
      tags: [Summaries]
      operationId: putProvinceGenerationSummaryNotAllowed
      summary: Not allowed. A generation summary is read-only (405)
      security: []
      responses:
        '405': { $ref: '#/components/responses/MethodNotAllowedSummary' }
    patch:
      tags: [Summaries]
      operationId: patchProvinceGenerationSummaryNotAllowed
      summary: Not allowed. A generation summary is read-only (405)
      security: []
      responses:
        '405': { $ref: '#/components/responses/MethodNotAllowedSummary' }
    delete:
      tags: [Summaries]
      operationId: deleteProvinceGenerationSummaryNotAllowed
      summary: Not allowed. A generation summary is read-only (405)
      security: []
      responses:
        '405': { $ref: '#/components/responses/MethodNotAllowedSummary' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT bearer token. Scopes (installation-write, analyst-read-by-jurisdiction, asset-admin) are documented in Step 9.9.
  parameters:
    ProvinceId:
      name: province-id
      in: path
      required: true
      schema: { type: string, pattern: '^PV-[0-9]{2}$', example: PV-01 }
...
    Date:
      name: date
      in: query
      description: Sri Lanka calendar day (Asia/Colombo) to summarise, YYYY-MM-DD. Defaults to today. An impossible date, another format or a day after today → 400 INVALID_QUERY_PARAMETER.
      schema: { type: string, format: date, example: '2026-10-06' }
    IfNoneMatch:
      name: If-None-Match
      in: header
      description: ETag(s) the client holds. A match returns 304.
      schema: { type: string, example: '"a3f9c2d1"' }
    IfModifiedSince:
      name: If-Modified-Since
      in: header
      description: Used only when If-None-Match is absent. Not modified since this date returns 304.
      schema: { type: string, example: 'Tue, 06 Oct 2026 04:30:00 GMT' }
...
  headers:
...
    ETag:
      description: Strong validator, a quoted hash of the response body
      schema: { type: string, example: '"a3f9c2d1"' }
    LastModified:
      description: When the representation last changed. Sent on single resources only, never on a collection or page.
      schema: { type: string, example: 'Tue, 06 Oct 2026 04:30:04 GMT' }
...
    Vary:
      schema: { type: string, example: 'Accept, Authorization' }
    CacheControl:
      schema: { type: string, example: 'private, no-cache' }
    WWWAuthenticate:
      schema: { type: string, example: 'Bearer realm="slsea-solar-api"' }
    AllowReading:
      description: The methods this URI does support
      schema: { type: string, example: 'GET, HEAD' }
...
  responses:
    NotModified:
      description: The client's copy is current. No body. Last-Modified is present only for single resources.
      headers:
        ETag: { $ref: '#/components/headers/ETag' }
        Last-Modified: { $ref: '#/components/headers/LastModified' }
        Vary: { $ref: '#/components/headers/Vary' }
        Cache-Control: { $ref: '#/components/headers/CacheControl' }
    BadRequest:
      description: Invalid body or query value
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: Missing, invalid or expired token
      headers:
        WWW-Authenticate: { $ref: '#/components/headers/WWWAuthenticate' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    Forbidden:
      description: Valid token but not allowed (scope, installation or jurisdiction)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: The member or its parent does not exist
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotAcceptable:
      description: The Accept header excludes application/json
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
    ProvinceNotFound:
      description: No province with this id (PROVINCE_NOT_FOUND)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: PROVINCE_NOT_FOUND, message: No province with id PV-99., details: [], more_info: /api-docs#error-codes }
...
    MethodNotAllowedSummary:
      description: METHOD_NOT_ALLOWED. A generation summary is worked out on request and is read-only; Allow is GET, HEAD.
      headers:
        Allow: { $ref: '#/components/headers/AllowReading' }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: METHOD_NOT_ALLOWED, message: 'A generation summary is read-only: POST is not allowed here. Allowed: GET, HEAD.', details: [], more_info: /api-docs#error-codes }
...
    ServiceUnavailable:
      description: The database cannot be reached (SERVICE_UNAVAILABLE)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
          example: { code: SERVICE_UNAVAILABLE, message: The database cannot be reached., details: [], more_info: /api-docs#error-codes }
  schemas:
...
    Province:
      type: object
      required: [province_id, name]
      properties:
        province_id: { type: string, example: PV-01 }
        name: { type: string, example: Western }
...
    ProvinceGenerationSummary:
      type: object
      description: The district summary one level up, from the same calculation; every field means the same.
      required: [province_id, date, total_power_kw, power_as_of, day_energy_kwh, peak_power_kw, peak_at, installed_capacity_kw, capacity_utilisation, installation_count, reporting_installations]
      properties:
        province_id: { type: string, example: PV-01 }
        date: { type: string, format: date, example: '2026-10-06' }
        total_power_kw: { type: number, example: 312.5 }
        power_as_of: { type: string, format: date-time, nullable: true, example: '2026-10-06T04:30:00Z' }
        day_energy_kwh: { type: number, example: 1410.2 }
        peak_power_kw: { type: number, example: 402.7 }
        peak_at: { type: string, format: date-time, nullable: true, example: '2026-10-06T06:30:00.000Z' }
        installed_capacity_kw: { type: number, example: 640 }
        capacity_utilisation: { type: number, nullable: true, example: 0.488 }
        installation_count: { type: integer, example: 95 }
        reporting_installations: { type: integer, example: 93 }
    ErrorDetail:
      type: object
      required: [field, location, issue, reference]
      properties:
        field: { type: string, nullable: true, example: power_kw }
        location: { type: string, nullable: true, enum: [path, query, body, header, null], example: body }
        issue: { type: string, example: Must be 0 or more. }
        reference: { type: string, nullable: true, example: null }
    Error:
      type: object
      required: [code, message, details, more_info]
      properties:
        code: { type: string, description: Stable machine-readable code (see Error codes), example: VALIDATION_FAILED }
        message: { type: string, example: The reading has 1 invalid field. }
        details:
          type: array
          items: { $ref: '#/components/schemas/ErrorDetail' }
        more_info: { type: string, example: '/api-docs#error-codes' }